+++ Kernel_23/doc/Kernel_23/Concepts/FunctionObjectConcepts.h
@@ -743,6 +743,33 @@ public:
 
 }; /* end Kernel::CompareAngleWithXAxis_2 */
 
+
+
+/*!
+  \ingroup PkgKernel23ConceptsFunctionObjects
+  \cgalConcept
+
+*/
+class CompareAngle_3 {
+public:
+
+  /// \name Operations
+  /// A model of this concept must provide:
+  /// @{
+
+  /*!
+  compares the angles \f$ \theta_1\f$ and \f$ \theta_2\f$, where
+  \f$ \theta_1\f$ is the angle, in \f$ [0, \pi]\f$, of the triangle
+  \f$ (a, b, c)\f$ at the vertex `b`, and \f$ \theta_2\f$ is
+  the angle in \f$ [0, \pi]\f$ such that \f$ cos(\theta_2) = cosine\f$.
+  \pre `a!=b && c!=b`.
+  */
+  Comparison_result operator()(const K::Point_3& a,
+                               const K::Point_3& b,
+                               const K::Point_3& c,
+                               const K::FT& cosine);
+};
+
 /*!
   \ingroup PkgKernel23ConceptsFunctionObjects
   \cgalConcept